Added min_fastlane_version. Made fastlane_version an alias.
